app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
pymongo==4.3.3
motor==3.1.1
pydantic==1.10.12
python-dotenv==1.0.1